% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximated_variance_estimators.R
\name{approx_var_est}
\alias{approx_var_est}
\title{Approximated Variance Estimators}
\usage{
approx_var_est(y, pik, method, sample = NULL, ...)
}
\arguments{
\item{y}{numeric vector of sample observations}

\item{pik}{numeric vector of first-order inclusion probabilities, of length *N* or *n*,
depending on the chosen method (see Details for more information)}

\item{method}{string indicating the desired approximate variance estimator}

\item{sample}{Either a numeric vector of length equal to the sample size, containing
the indeces of sample units, or a boolean vector of the same length of \code{pik}, indicating which
units belong to the sample (\code{TRUE} if the unit is in the sample,
\code{FALSE} otherwise.
Only used with estimators of the third class (see Details for more information).}

\item{...}{two optional parameters can be modified to control the iterative
procedures used \code{method} is \code{"MateiTille5"}, \code{"Tille"} or
\code{"FixedPoint"}: \code{maxIter} sets the maximum number
of iteration to perform and \code{eps} controls the convergence error}
}
\value{
a scalar, the estimated variance
}
\description{
Approximated variance estimators which use only first-order inclusion probabilities
}
\details{
Matei and Tillé (2005) define three classes:

1) first and second-order inclusion probabilities
The first class is composed of the Horvitz-Thompson estimator (Horvitz and Thompson 1952) and the Sen-Yates-Grundy estimator
(Yates and Grundy 1953; Sen 1953) and is not available in this package.
For this estimators, the use may refer to package \code{sampling} ...

2) only first-order inclusion probabilities, only for sample units

>>> TBD: add estimators names and formulas

3) only first-order inclusion probabilities, for the entire population

>>> TBD: add estimators names and formulas
}
\examples{

### Generate population data ---
N <- 500; n <- 50

set.seed(0)
x <- rgamma(500, scale=10, shape=5)
y <- abs( 2*x + 3.7*sqrt(x) * rnorm(N) )

pik <- n * x/sum(x)
s   <- sample(N, n)

ys <- y[s]
piks <- pik[s]

### Estimators of class 2 ---
approx_var_est(ys, piks, method="Deville1")
approx_var_est(ys, piks, method="Deville2")
approx_var_est(ys, piks, method="Deville3")
approx_var_est(ys, piks, method="Rosen")
approx_var_est(ys, piks, method="FixedPoint")
approx_var_est(ys, piks, method="Brewer2")

### Estimators of class 3 ---
approx_var_est(ys, pik, method="Berger", sample=s)
approx_var_est(ys, pik, method="Tille", sample=s)
approx_var_est(ys, pik, method="MateiTille1", sample=s)
approx_var_est(ys, pik, method="MateiTille2", sample=s)
approx_var_est(ys, pik, method="MateiTille3", sample=s)
approx_var_est(ys, pik, method="MateiTille4", sample=s)
approx_var_est(ys, pik, method="MateiTille5", sample=s)
approx_var_est(ys, pik, method="Brewer1", sample=s)
approx_var_est(ys, pik, method="Brewer3", sample=s)
approx_var_est(ys, pik, method="Brewer4", sample=s)


}
\references{
Matei, A.; Tillé, Y., 2005. Evaluation of variance approximations and estimators
in maximum entropy sampling with unequal probability and fixed sample size.
Journal of Official Statistics 21 (4), 543-570.

Haziza, D.; Mecatti, F.; Rao, J.N.K. 2008.
Evaluation of some approximate variance estimators under the Rao-Sampford
unequal probability sampling design. Metron LXVI (1), 91-108.
}
