% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTvar.R
\name{HTvar}
\alias{HTvar}
\title{Variance of the Horvitz-Thompson estimator}
\usage{
HTvar(y, pikl, sample = TRUE, method)
}
\arguments{
\item{y}{numeric vector representing the variable object of inference}

\item{pikl}{matrix of second-order (joint) inclusion probabilities; the diagonal
must contain the first-order inclusion probabilities.}

\item{sample}{boolean indicating if sample values are provided and, thus, an
estimate of the variance is required}

\item{method}{string, indicating if the Horvitz-Thompson (\code{"HT"}) or the
Sen-Yates-Grundy (\code{"SYG"}) estimator should be computed.}
}
\description{
Compute or estimate the variance of the Horvitz-Thompson total estimator
by the Horvitz-Thompson or Sen-Yates-Grundy variance estimators.
}
