% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-UPSvarApprox.R
\docType{package}
\name{UPSvarApprox}
\alias{UPSvarApprox}
\alias{UPSvarApprox-package}
\title{UPSvarApprox: Approximate the variance of the Horvitz-thompson estimator}
\description{
UPSvarApprox: Approximate the variance of the Horvitz-thompson estimator
}
\section{Variance approximation}{

The package provides function \code{Var_approx} for the approximation of the
Horvitz-Thompson variance, and function \code{approx_var_est} for the computation
of approximate variance estimators.
For both functions, different estimators are implemented,
see their documentation for details.
}

\references{
Matei, A.; Tillé, Y., 2005. Evaluation of variance approximations and estimators
in maximum entropy sampling with unequal probability and fixed sample size.
Journal of Official Statistics 21 (4), 543-570.

Haziza, D.; Mecatti, F.; Rao, J.N.K. 2008.
Evaluation of some approximate variance estimators under the Rao-Sampford
unequal probability sampling design. Metron LXVI (1), 91-108.
}
