% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Variance_approximations.R
\name{Var_approx}
\alias{Var_approx}
\title{Approximate the Variance of the Horvitz-Thompson estimator}
\usage{
Var_approx(y, pik, n, method, ...)
}
\arguments{
\item{y}{numeric vector containing the values of the variable of interest, of
length equal to population size}

\item{pik}{numeric vector of first-order inclusion probabilities, of length
equal to population size}

\item{n}{a scalar indicating the sample size}

\item{method}{string indicating the approximation that should be used}

\item{...}{two optional parameters can be modified to control the iterative
procedure in \code{method="FixedPoint"}: \code{maxIter} sets the maximum number
of iteration to perform and \code{eps} controls the convergence error}
}
\value{
a scalar, the approximated variance.
}
\description{
Approximations of the Horvitz-Thompson variance for High-Entropy sampling designs.
Such methods use only first-order inclusion probabilities.
}
\details{
TBD
}
\examples{

N <- 500; n <- 50

set.seed(0)
x <- rgamma(n=N, scale=10, shape=5)
y <- abs( 2*x + 3.7*sqrt(x) * rnorm(N) )

pik  <- n * x/sum(x)
pikl <- outer(pik, pik, '*'); diag(pikl) <- pik

### Variance approximations ---
Var_approx(y, pik, n, method = "Hajek1")
Var_approx(y, pik, n, method = "Hajek1")
Var_approx(y, pik, n, method = "HartleyRao1")
Var_approx(y, pik, n, method = "HartleyRao2")
Var_approx(y, pik, n, method = "FixedPoint")


}
\references{
Matei, A.; Tillé, Y., 2005. Evaluation of variance approximations and estimators
in maximum entropy sampling with unequal probability and fixed sample size.
Journal of Official Statistics 21 (4), 543-570.
}
