% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varEst_class2.R
\name{var_Brewer_class2}
\alias{var_Brewer_class2}
\title{Approximate Variance Estimators by Brewer (2002)}
\usage{
var_Brewer_class2(y, pik)
}
\arguments{
\item{y}{numeric vector of sample observations}

\item{pik}{numeric vector of first-order inclusion probabilities for sample units}
}
\value{
a scalar, the estimated variance
}
\description{
Computes an approximate variance estimate according to one  of
the estimators proposed by Brewer (2002).
This estimator belongs to class 2, it requires only first-order inclusion
probabilities and only for sample units.
}
\keyword{internal}
