% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varEst_class2.R
\name{var_Deville}
\alias{var_Deville}
\title{Deville's approximate variance estimators}
\usage{
var_Deville(y, pik, method)
}
\arguments{
\item{y}{numeric vector of sample observations}

\item{pik}{numeric vector of first-order inclusion probabilities for sample units}

\item{method}{string indicating the desired approximate variance estimator.
One of "Deville1", "Deville2", "Deville3", "Hajek", "Rosen", "FixedPoint",
"Brewer1", "HartleyRao", "Berger", "Tille", "MateiTille1", "MateiTille2",
"MateiTille3", "MateiTille4", "MateiTille5", "Brewer2", "Brewer3", "Brewer4".}
}
\value{
a scalar, the estimated variance
}
\description{
Compute Deville's approximate variance estimators of class 2, which require only
first-order inclusion probabilities and only for sample units
}
