% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varEst_class2.R
\name{var_FixedPoint}
\alias{var_FixedPoint}
\title{Fixed-Point approximate variance estimator}
\usage{
var_FixedPoint(y, pik, maxIter = 1000, eps = 1e-05)
}
\arguments{
\item{y}{numeric vector of sample observations}

\item{pik}{numeric vector of first-order inclusion probabilities for sample units}

\item{maxIter}{a scalar indicating the maximum number of iterations for the
fixed-point procedure}

\item{eps}{tolerance value for the convergence of the fixed-point procedure}
}
\value{
a scalar, the estimated variance
}
\description{
Fixed-Point estimator for approximate variance estimation by Deville and Tillé (2005).
Estimator of class 2, it requires only first-order inclusion probabilities and
only for sample units.
}
