% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varEst_class3.R
\name{var_HartleyRao}
\alias{var_HartleyRao}
\title{Hartley and Rao approximate variance estimator}
\usage{
var_HartleyRao(y, pik, sample)
}
\arguments{
\item{y}{numeric vector of sample observations}

\item{pik}{numeric vector of first-order inclusion probabilities for all population units}

\item{sample}{Either a numeric vector of length equal to the sample size with
the indices of sample units, or a boolean vector of the same length of \code{pik}, indicating which
units belong to the sample (\code{TRUE} if the unit is in the sample,
\code{FALSE} otherwise.
Only used with estimators of the third class (see Details for more information).}
}
\value{
a scalar, the estimated variance
}
\description{
Compute the an approximate variance estimator obtained by the approximation
of joint-inclusion probabilities proposed by Hartley and Rao (1962).
Estimator of class 3, it requires only first-order inclusion probabilities but
for all population units.
}
\keyword{internal}
