% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varEst_class2.R
\name{var_Rosen}
\alias{var_Rosen}
\title{Rosén Approximate Variance Estimator}
\usage{
var_Rosen(y, pik)
}
\arguments{
\item{y}{numeric vector of sample observations}

\item{pik}{numeric vector of first-order inclusion probabilities for sample units}
}
\value{
a scalar, the estimated variance
}
\description{
Compute an approximate variance estimate using the estimator
proposed by Rosén (1991).
Estimator of class 2, it requires only first-order inclusion probabilities and
only for sample units.
}
