% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jip_approximations.R
\name{jip_approx}
\alias{jip_approx}
\title{Approximate Joint-Inclusion Probabilities}
\usage{
jip_approx(pik, method)
}
\arguments{
\item{pik}{numeric, vector of first-order inclusion probabilities for all
population units.}

\item{method}{string representing one of the available approximation methods.}
}
\value{
A symmetric matrix of inclusion probabilities, which diagonal is the
vector of first-order inclusion probabilities.
}
\description{
Approximations of joint-inclusion probabilities by means of first-order
inclusion probabilities.
}
